factor out config loading into separate module
